add feat_rbac

+++ src/models/users.model.ts
@@ -1,17 +1,18 @@
 import { ResultSetHeader } from 'mysql2/promise'
 import pool from '../config/database.config';
 import CustomError from '../service/customError.service';
-import { IUser, IUserGetInfo, IUserPostInfo } from '../types/user.interface';
+import { IUser, IUserGetInfo, IUserPostInfo } from '../types/user.data';
 import { errorInfo } from '../service/handleError.service';
+import { StatusCodes } from 'http-status-codes';
 
 class UsersModel {
     async getUsers() {
         try{
             const connection = await pool.getConnection();
-            const [rows] = await connection.query<IUserGetInfo[]>('SELECT Email, CAST(Gender AS VARCHAR(255)) AS Gender, Age, CAST(Role AS VARCHAR(255)) AS Role FROM users');
+            const [rows] = await connection.query<IUserGetInfo[]>('SELECT Email, CAST(Gender AS VARCHAR(255)) AS Gender, Age FROM users');
             connection.release();
             return rows;
-        }catch(error: unknown){
+        }catch(error){
             const { statusError, messageError } = errorInfo(error);
             throw new CustomError(statusError, messageError);
         }
@@ -20,12 +21,16 @@ class UsersModel {
     async getDetailUser(userId: number){
         try{
             const connection = await pool.getConnection();
-            const query = `SELECT Email, CAST (Gender AS VARCHAR(255)) AS Gender, Age, CAST (Role AS VARCHAR(255)) AS Role FROM users WHERE UserID = ?;`;
-            const value = [userId];
-            const [rows] = await connection.query<IUserGetInfo[]>(query, value);
+            const query = `SELECT Email, CAST (Gender AS VARCHAR(255)) AS Gender, Age FROM users WHERE UserID = ?;`;
+            const [rows] = await connection.query<IUserGetInfo[]>(query, userId);
+
+            if(rows[0] == null){
+                throw new CustomError(StatusCodes.NOT_FOUND, 'User not found'); //404
+            }
+
             connection.release();
             return rows[0];
-        }catch(error: unknown){
+        }catch(error){
             const { statusError, messageError } = errorInfo(error);
             throw new CustomError(statusError, messageError);
         }
@@ -33,14 +38,25 @@ class UsersModel {
 
     async createUser(user: IUserPostInfo){
         try{
-            const connection = await pool.getConnection();
-            const query = `INSERT INTO users (Email, Password, Gender, Age, Role) VALUES (?, ?, ?, ?, ?);`;
             const {Email, Password, Gender, Age, Role} = user;
-            const value = [Email, Password, Gender, Age, Role];
-            const results = await connection.query<ResultSetHeader>(query, value);
+
+            const connection = await pool.getConnection();
+            
+            //insert users
+            const queryUser = `INSERT INTO users (Email, Password, Gender, Age) VALUES (?, ?, ?, ?);`;
+            const valueUser = [Email, Password, Gender, Age];
+            const resultsUser = await connection.query<ResultSetHeader>(queryUser, valueUser);
+
+            //insert user-role
+            for(let i = 0; i < Role.length; i++){
+                const queryRole = `INSERT INTO users_roles (UserID, RoleID) VALUES (?, ?);`;
+                const valueRole = [resultsUser[0].insertId, Role[i]];
+                await connection.query<ResultSetHeader>(queryRole, valueRole);
+            }
+
             connection.release();
-            return results[0].insertId;
-        }catch(error: unknown){
+            return resultsUser[0].insertId;
+        }catch(error){
             const { statusError, messageError } = errorInfo(error);
             throw new CustomError(statusError, messageError);
         }
@@ -48,14 +64,17 @@ class UsersModel {
 
     async updateUser(user: IUserPostInfo, userID: number){
         try{
-            const connection = await pool.getConnection();
-            const query = `UPDATE users SET Email = ?, Password = ?, Gender = ?, Age = ?, Role = ? WHERE UserID = ?`;
             const {Email, Password, Gender, Age, Role} = user;
-            const value = [Email, Password, Gender, Age, Role, userID];
+
+            const connection = await pool.getConnection();
+
+            const query = `UPDATE users SET Email = ?, Password = ?, Gender = ?, Age = ? WHERE UserID = ?`;
+            const value = [Email, Password, Gender, Age, userID];
             const results = await connection.query<ResultSetHeader>(query, value);
+
             connection.release();
             return results[0].affectedRows;
-        }catch(error: unknown){
+        }catch(error){
             const { statusError, messageError } = errorInfo(error);
             throw new CustomError(statusError, messageError);
         }
@@ -64,12 +83,16 @@ class UsersModel {
     async deleteUser(userId: number){
         try{
             const connection = await pool.getConnection();
-            const query = `DELETE FROM users WHERE UserID = ?`;
-            const value = [userId];
-            const results = await connection.query<ResultSetHeader>(query, value);
+
+            //delete role of user
+            await connection.query<ResultSetHeader>('DELETE FROM users_roles WHERE UserID = ?', userId);
+
+            //delete user
+            const resultsUser = await connection.query<ResultSetHeader>('DELETE FROM users WHERE UserID = ?', userId);
+
             connection.release();
-            return results[0].affectedRows;
-        }catch(error: unknown){
+            return resultsUser[0].affectedRows;
+        }catch(error){
             const { statusError, messageError } = errorInfo(error);
             throw new CustomError(statusError, messageError);
         }
@@ -78,13 +101,11 @@ class UsersModel {
     async getUserByEmail(email: string) {
         try{
             const connection = await pool.getConnection();
-            const query = `SELECT * FROM users WHERE Email = ?`;
-            const value = [email];
-            const [rows] = await connection.query<IUser[]>(query, value);
+            const [rows] = await connection.query<IUser[]>('SELECT * FROM users WHERE Email = ?', email);
             connection.release();
             console.log(rows[0])
             return rows[0];
-        }catch(error: unknown){
+        }catch(error){
             const { statusError, messageError } = errorInfo(error);
             throw new CustomError(statusError, messageError);
         }
@@ -93,12 +114,10 @@ class UsersModel {
     async setAccessToken(accessToken: string, email: string){
         try{
             const connection = await pool.getConnection();
-            const query = `UPDATE users SET AccessToken = ? WHERE Email = ?`;
-            const value = [accessToken, email];
-            await connection.query(query, value);
+            await connection.query('UPDATE users SET AccessToken = ? WHERE Email = ?', [accessToken, email]);
             connection.release();
             return true;
-        }catch(error: unknown){
+        }catch(error){
             const { statusError, messageError } = errorInfo(error);
             throw new CustomError(statusError, messageError);
         }
@@ -107,12 +126,10 @@ class UsersModel {
     async getAccessTokenByUserID(userId: number){
         try{
             const connection = await pool.getConnection();
-            const query = `SELECT * FROM users WHERE UserID = ?`;
-            const value = [userId];
-            const [rows] = await connection.query<IUser[]>(query, value);
+            const [rows] = await connection.query<IUser[]>('SELECT * FROM users WHERE UserID = ?', userId);
             connection.release();
             return rows[0].AccessToken;
-        }catch(error: unknown){
+        }catch(error){
             const { statusError, messageError } = errorInfo(error);
             throw new CustomError(statusError, messageError);
         }
